Add runtime-specific TypeScript SDK exports
