+++ packages/sdk-ts/src/index.shared.ts
@@ -1,0 +1,21 @@
+export {
+  BrowserContext,
+  type ClearCookieOptions,
+  type Cookie,
+  type CookieParam,
+  type DomainPolicy,
+} from "./browserContext.js";
+export {
+  BrowserClipboard,
+  type ClipboardOptions,
+  type ClipboardPasteOptions,
+} from "./browserClipboard.js";
+export { Locator } from "./locator.js";
+export { Page, type ScreenshotOptions } from "./page.js";
+export type { InitScriptSource } from "./pageScripts.js";
+export type {
+  BrowserGetVersionResult,
+  RuntimeLoopbackStatusResult,
+  StagehandMetrics,
+  StagehandPingResult,
+} from "../../protocol/types.js";
