+++ packages/sdk-ts/src/index.ts
@@ -1,39 +1,3 @@
-export {
-  BrowserContext,
-  type ClearCookieOptions,
-  type Cookie,
-  type CookieParam,
-  type DomainPolicy,
-} from "./browserContext.js";
-export {
-  BrowserClipboard,
-  type ClipboardOptions,
-  type ClipboardPasteOptions,
-} from "./browserClipboard.js";
-export { Locator } from "./locator.js";
-export { Page, type ScreenshotOptions } from "./page.js";
-export type { InitScriptSource } from "./pageScripts.js";
-export { Stagehand } from "./stagehand.js";
-export type {
-  BrowserGetVersionResult,
-  RuntimeLoopbackStatusResult,
-  StagehandMetrics,
-  StagehandPingResult,
-} from "../../protocol/types.js";
-export {
-  BrowserSourceSchema,
-  BrowserbaseBrowserSourceSchema,
-  CdpBrowserSourceSchema,
-  ClientLLMSchema,
-  LocalBrowserSourceSchema,
-  StagehandClientLogFormatSchema,
-  StagehandClientLoggingConfigSchema,
-  StagehandClientLogLevelSchema,
-  StagehandClientInitParamsSchema,
-  type BrowserSource,
-  type ClientLLM,
-  type ResolvedStagehandClientLoggingConfig,
-  type ResolvedStagehandClientInitParams,
-  type StagehandClientLoggingConfig,
-  type StagehandClientInitParams,
-} from "./clientSchemas.js";
+// Source imports and repository tests use the Node-capable variant. Published consumers are routed
+// to the Node or web artifact by the package export conditions.
+export * from "./runtime/node/index.js";
