+++ packages/sdk-ts/src/runtime/node/clientSchemas.ts
@@ -1,0 +1,22 @@
+import { z } from "zod/v4";
+import {
+  BrowserbaseBrowserSourceSchema,
+  CdpBrowserSourceSchema,
+  createStagehandClientInitParamsSchema,
+  LocalBrowserSourceSchema,
+} from "../../clientSchemas.shared.js";
+
+export const BrowserSourceSchema = z
+  .discriminatedUnion("type", [
+    BrowserbaseBrowserSourceSchema,
+    LocalBrowserSourceSchema,
+    CdpBrowserSourceSchema,
+  ])
+  .meta({ id: "BrowserSource" });
+
+export const StagehandClientInitParamsSchema =
+  createStagehandClientInitParamsSchema(BrowserSourceSchema);
+
+export type BrowserSource = z.infer<typeof BrowserSourceSchema>;
+export type StagehandClientInitParams = z.input<typeof StagehandClientInitParamsSchema>;
+export type ResolvedStagehandClientInitParams = z.output<typeof StagehandClientInitParamsSchema>;
