+++ packages/sdk-ts/src/runtime/web/index.ts
@@ -1,0 +1,21 @@
+export * from "../../index.shared.js";
+export { Stagehand } from "./stagehand.js";
+export {
+  BrowserSourceSchema,
+  StagehandClientInitParamsSchema,
+  type BrowserSource,
+  type ResolvedStagehandClientInitParams,
+  type StagehandClientInitParams,
+} from "./clientSchemas.js";
+export {
+  BrowserbaseBrowserSourceSchema,
+  CdpBrowserSourceSchema,
+  ClientLLMSchema,
+  LocalBrowserSourceSchema,
+  StagehandClientLogFormatSchema,
+  StagehandClientLoggingConfigSchema,
+  StagehandClientLogLevelSchema,
+  type ClientLLM,
+  type ResolvedStagehandClientLoggingConfig,
+  type StagehandClientLoggingConfig,
+} from "../../clientSchemas.shared.js";
